refactor: use aiometer to improve performance
