+++ abuse_whois/__init__.py
@@ -1,18 +1,4 @@
-from typing import Optional
-
-from abuse_whois.matchers.shared_hosting import get_shared_hosting_provider
-from abuse_whois.matchers.whois import get_contact_from_whois
-
-from .errors import InvalidAddressError
-from .ip import resolve_ip_address
-from .schemas import Contact, Contacts
-from .utils import (
-    get_hostname,
-    get_registered_domain,
-    is_domain,
-    is_ip_address,
-    is_supported_address,
-)
+from .main import get_abuse_contacts  # noqa: F401
 
 try:
     import importlib.metadata as importlib_metadata
@@ -21,44 +7,3 @@ except ModuleNotFoundError:
 
 
 __version__ = importlib_metadata.version(__name__)
-
-
-async def get_abuse_contacts(address: str) -> Contacts:
-    if not is_supported_address(address):
-        raise InvalidAddressError(f"{address} is not supported type address")
-
-    shared_hosting_provider: Optional[Contact] = None
-    registrar: Optional[Contact] = None
-    hosting_provider: Optional[Contact] = None
-
-    hostname = get_hostname(address)
-    ip_address: Optional[str] = None
-    registered_domain: Optional[str] = None
-
-    shared_hosting_provider = get_shared_hosting_provider(hostname)
-
-    if is_domain(hostname):
-        registered_domain = get_registered_domain(hostname)
-        registrar = await get_contact_from_whois(hostname)
-
-        # get IP address by domain
-        try:
-            ip_address = await resolve_ip_address(hostname)
-        except OSError:
-            pass
-
-    if is_ip_address(hostname):
-        ip_address = hostname
-
-    if ip_address is not None:
-        hosting_provider = await get_contact_from_whois(ip_address)
-
-    return Contacts(
-        address=address,
-        hostname=hostname,
-        ip_address=ip_address,
-        registered_domain=registered_domain,
-        shared_hosting_provider=shared_hosting_provider,
-        registrar=registrar,
-        hosting_provider=hosting_provider,
-    )
