+++ pyproject.toml
@@ -10,21 +10,22 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.7"
+aiometer = "^0.3.0"
 asyncer = "^0.0.1"
 cachetools = "^5.2.0"
 email-validator = "^1.2.1"
 fastapi = "^0.78.0"
 gunicorn = { version = "^20.1.0", optional = true }
 importlib-metadata = {version = "^4.11.4", python = "<3.8"}
-typing-extensions = {version = "^4.2.0", python = "<3.8"}
 loguru = "^0.6.0"
-pydantic = "^1.9.0"
-pyhumps = "^3.7.1"
+pydantic = "^1.9.1"
+pyhumps = "^3.7.2"
 PyYAML = "^6.0"
 sh = "^1.14.2"
 tldextract = "^3.3.0"
 typer = "^0.4.1"
-uvicorn = { extras = ["standard"], version = "^0.17.6", optional = true }
+typing-extensions = {version = "^3.10", python = "<3.8"}
+uvicorn = { extras = ["standard"], version = "^0.18.1", optional = true }
 whois-parser = "^0.1.4"
 
 [tool.poetry.dev-dependencies]
