+++ tests/terst_version.py
@@ -1,0 +1,5 @@
+from abuse_whois import __version__
+
+
+def test_version():
+    assert isinstance(__version__, str)
